lift_hardware_controller.py
try:
    import serial
    SERIAL_AVAILABLE = True
except ImportError:
    print("[WARN] pyserial not installed. Hardware control will use simulation mode.")
    SERIAL_AVAILABLE = False

import time

class HardwareController:
    """Phase 3 Production Edge Hardware Controller"""
    def __init__(self, port='COM3', baudrate=9600, simulate=True):
        self.simulate = simulate or not SERIAL_AVAILABLE
        self.arduino = None
        
        if not self.simulate and SERIAL_AVAILABLE:
            try:
                # Production Serial Communication to Relay / PLC Unit
                self.arduino = serial.Serial(port, baudrate, timeout=1)
                time.sleep(2) # Handshake stabilization
                print(f"[Hardware] Synced securely via RS485/Serial on {port}")
            except Exception as e:
                print(f"[!] Hardware Comms Failure: {e}")
                print("[!] Bypassing to Local Simulation Mode...")
                self.simulate = True
        else:
            if not SERIAL_AVAILABLE:
                print("[Hardware] Running in simulation mode (pyserial not installed)")
            else:
                print("[Hardware] Running in simulation mode (simulate=True)")

    def dispatch_lift(self, floor):
        """Electronically triggers a floor dispatch via High-Voltage relays"""
        # We enforce a phantom 500ms button press directly on the physical line
        if self.simulate:
            print(f"\n======================================")
            print(f">>> [HW-CORE]: ENERGIZING RELAYS")
            print(f">>> [HW-CORE]: LIFT DISPATCHED TO FLOOR {floor}")
            print(f"======================================\n")
            return True
            
        try:
            # Physical RS485 / UART transmission protocol
            command = f"GOTO:{floor}\n".encode()
            if self.arduino:
                self.arduino.write(command)
            return True
        except Exception as e:
            print(f"[HW-CRITICAL-ERROR]: Serial transmission failed: {e}")
            return False
