requirements.txt
# SmartLift Requirements
# Core Web Framework
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3

# Computer Vision & Face Recognition (FAST VERSION)
# Using face-recognition instead of DeepFace for 10x speed improvement
opencv-python==4.10.0.84
face-recognition==1.3.0
dlib==19.24.6
Pillow==11.0.0

# Audio Processing
pyttsx3==2.98
SpeechRecognition==3.14.0
PyAudio==0.2.14

# QR Code Generation
qrcode==8.0
python-qrcode[pil]==8.0

# Hardware Communication (optional for demo mode)
pyserial==3.5

# Email Support
# Uses built-in smtplib, no extra deps needed

# Database
# SQLite is built-in with Python

# Additional utilities
numpy==2.2.5
